<!-- Describe what has changed in this PR -->
**What changed?**
* Optimize otel metrics filtering

<!-- Tell your future self why have you made these changes -->
**Why?**
```
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkTagsToAttributes_Old_50-16    	 4428312	       237.1 ns/op	     192 B/op	       2 allocs/op
BenchmarkTagsToAttributes_New_50-16    	10890871	       197.3 ns/op	     128 B/op	       1 allocs/op
BenchmarkTagsToAttributes_Old_00-16    	 7385574	       160.6 ns/op	     192 B/op	       2 allocs/op
BenchmarkTagsToAttributes_New_00-16    	12174460	       141.2 ns/op	     128 B/op	       1 allocs/op
```

<!-- How have you verified this change? Tested locally? Added a unit test? Checked in staging env? -->
**How did you test it?**
N/A

<!-- Assuming the worst case, what can be broken when deploying this change to production? -->
**Potential risks**
N/A

<!-- Is this PR a hotfix candidate or require that a notification be sent to the broader community? (Yes/No) -->
**Is hotfix candidate?**
N/A