@wxing1292 2022-11-10
```
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkTagsToAttributes_Old_50-16    	 4883530	       268.2 ns/op	     192 B/op	       2 allocs/op
BenchmarkTagsToAttributes_New_50-16    	 9390368	       174.6 ns/op	     128 B/op	       1 allocs/op
BenchmarkTagsToAttributes_Old_00-16    	 7229772	       162.6 ns/op	     192 B/op	       2 allocs/op
BenchmarkTagsToAttributes_New_00-16    	11213854	       134.9 ns/op	     128 B/op	       1 allocs/op
```

```
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkTagsToAttributes_Old_50-16    	 5420859	       235.8 ns/op	     192 B/op	       2 allocs/op
BenchmarkTagsToAttributes_New_50-16    	 9328764	       170.2 ns/op	     128 B/op	       1 allocs/op
BenchmarkTagsToAttributes_Old_00-16    	 7381803	       185.2 ns/op	     192 B/op	       2 allocs/op
BenchmarkTagsToAttributes_New_00-16    	13326768	       166.7 ns/op	     128 B/op	       1 allocs/op
```